+++ R/parse-iupac-condensed.R
@@ -35,10 +35,13 @@
 #' @seealso [parse_iupac_short()], [parse_iupac_extended()]
 #'
 #' @export
-parse_iupac_condensed <- function(x, on_failure = "error", progress = FALSE) {
-  struc_parser_wrapper(
+parse_iupac_condensed <- function(
+  x,
+  on_failure = "error",
+  progress = FALSE
+) {
+  normalized_struc_parser_wrapper(
     x,
-    do_parse_iupac_condensed,
     on_failure = on_failure,
     progress = progress
   )