+++ R/parse-linear-code.R
@@ -15,10 +15,14 @@
 #' parse_linear_code(linear_code)
 #'
 #' @export
-parse_linear_code <- function(x, on_failure = "error", progress = FALSE) {
-  struc_parser_wrapper(
+parse_linear_code <- function(
+  x,
+  on_failure = "error",
+  progress = FALSE
+) {
+  normalized_struc_parser_wrapper(
     x,
-    do_parse_linear_code,
+    convert_linear_to_iupac,
     on_failure = on_failure,
     progress = progress
   )